Upgrade serde 0.7.8 -> 1.0.8 and use macros 1.1
secrets' `api.rs` uses serde custom `derive()` to generate the json generation
and parsing code for API objects sent between the client and server.

Prior to rust 1.15.0, serde's procedural macro based custom `derive()` required
a nightly compiler. Using stable rust the technique available was to use a
`build.rs` to generate `api.rs` out of `api.rs.in` and use some `include!` magic
to try to make that a little more transparent. With 1.15,
[Macros 1.1](https://github.com/rust-lang/rfcs/blob/master/text/1681-macros-1.1.md)
was released making this process much more ergonomic and no longer requires the
build-time magic.

This commit upgrades us to use that technique by upgrading serde from 0.7.8 to
1.0.8 which also requires pulling in serde_macros and an upgrade to serde_json
and sodiumoxide. This removes the requirement for `build.rs` and allows for a
simple rename of `api.rs.in` to `api.rs`